+++ src/torchmetrics/utilities/__init__.py
@@ -13,15 +13,16 @@
 # limitations under the License.
 from torchmetrics.utilities.checks import check_forward_full_state_property
 from torchmetrics.utilities.data import apply_to_collection
-from torchmetrics.utilities.distributed import class_reduce, reduce
+from torchmetrics.utilities.distributed import EvaluationDistributedSampler, class_reduce, reduce
 from torchmetrics.utilities.prints import rank_zero_debug, rank_zero_info, rank_zero_warn
 
 __all__ = [
-    "check_forward_full_state_property",
     "apply_to_collection",
+    "check_forward_full_state_property",
     "class_reduce",
-    "reduce",
+    "EvaluationDistributedSampler",
     "rank_zero_debug",
     "rank_zero_info",
     "rank_zero_warn",
+    "reduce",
 ]